+++ .vim_runtime/my_plugins/fzf/src/terminal.go
@@ -4,7 +4,6 @@ import (
 	"bufio"
 	"bytes"
 	"fmt"
-	"io"
 	"io/ioutil"
 	"os"
 	"os/signal"
@@ -23,12 +22,14 @@ import (
 // import "github.com/pkg/profile"
 
 var placeholder *regexp.Regexp
+var numericPrefix *regexp.Regexp
 var activeTempFiles []string
 
 const ellipsis string = ".."
 
 func init() {
 	placeholder = regexp.MustCompile(`\\?(?:{[+sf]*[0-9,-.]*}|{q}|{\+?f?nf?})`)
+	numericPrefix = regexp.MustCompile(`^[[:punct:]]*([0-9]+)`)
 	activeTempFiles = []string{}
 }
 
@@ -41,11 +42,25 @@ const (
 )
 
 type previewer struct {
-	text    string
-	lines   int
-	offset  int
-	enabled bool
-	more    bool
+	version    int
+	lines      []string
+	offset     int
+	enabled    bool
+	scrollable bool
+	final      bool
+	spinner    string
+}
+
+type previewed struct {
+	version  int
+	numLines int
+	offset   int
+	filled   bool
+}
+
+type eachLine struct {
+	line string
+	err  error
 }
 
 type itemLine struct {
@@ -64,7 +79,7 @@ type Terminal struct {
 	initDelay    time.Duration
 	infoStyle    infoStyle
 	spinner      []string
-	prompt       string
+	prompt       func()
 	promptLen    int
 	pointer      string
 	pointerLen   int
@@ -123,6 +138,7 @@ type Terminal struct {
 	reqBox       *util.EventBox
 	preview      previewOpts
 	previewer    previewer
+	previewed    previewed
 	previewBox   *util.EventBox
 	eventBox     *util.EventBox
 	mutex        sync.Mutex
@@ -169,6 +185,7 @@ const (
 	reqPreviewEnqueue
 	reqPreviewDisplay
 	reqPreviewRefresh
+	reqPreviewDelayed
 	reqQuit
 )
 
@@ -224,14 +241,18 @@ const (
 	actJump
 	actJumpAccept
 	actPrintQuery
+	actRefreshPreview
 	actReplaceQuery
 	actToggleSort
 	actTogglePreview
 	actTogglePreviewWrap
+	actPreview
 	actPreviewUp
 	actPreviewDown
 	actPreviewPageUp
 	actPreviewPageDown
+	actPreviewHalfPageUp
+	actPreviewHalfPageDown
 	actPreviousHistory
 	actNextHistory
 	actExecute
@@ -253,6 +274,19 @@ type placeholderFlags struct {
 type searchRequest struct {
 	sort    bool
 	command *string
+}
+
+type previewRequest struct {
+	template string
+	pwindow  tui.Window
+	list     []*Item
+}
+
+type previewResult struct {
+	version int
+	lines   []string
+	offset  int
+	spinner string
 }
 
 func toActions(types ...actionType) []action {
@@ -326,6 +360,24 @@ func trimQuery(query string) []rune {
 	return []rune(strings.Replace(query, "\t", " ", -1))
 }
 
+func hasPreviewAction(opts *Options) bool {
+	for _, actions := range opts.Keymap {
+		for _, action := range actions {
+			if action.t == actPreview {
+				return true
+			}
+		}
+	}
+	return false
+}
+
+func makeSpinner(unicode bool) []string {
+	if unicode {
+		return []string{`⠋`, `⠙`, `⠹`, `⠸`, `⠼`, `⠴`, `⠦`, `⠧`, `⠇`, `⠏`}
+	}
+	return []string{`-`, `\`, `|`, `/`, `-`, `\`, `|`, `/`}
+}
+
 // NewTerminal returns new Terminal object
 func NewTerminal(opts *Options, eventBox *util.EventBox) *Terminal {
 	input := trimQuery(opts.Query)
@@ -343,7 +395,7 @@ func NewTerminal(opts *Options, eventBox *util.EventBox) *Terminal {
 		delay = initialDelay
 	}
 	var previewBox *util.EventBox
-	if len(opts.Preview.command) > 0 {
+	if len(opts.Preview.command) > 0 || hasPreviewAction(opts) {
 		previewBox = util.NewEventBox()
 	}
 	strongAttr := tui.Bold
@@ -382,21 +434,17 @@ func NewTerminal(opts *Options, eventBox *util.EventBox) *Terminal {
 		}
 		renderer = tui.NewLightRenderer(opts.Theme, opts.Black, opts.Mouse, opts.Tabstop, opts.ClearOnExit, false, maxHeightFunc)
 	}
-	wordRubout := "[^[:alnum:]][[:alnum:]]"
-	wordNext := "[[:alnum:]][^[:alnum:]]|(.$)"
+	wordRubout := "[^\\pL\\pN][\\pL\\pN]"
+	wordNext := "[\\pL\\pN][^\\pL\\pN]|(.$)"
 	if opts.FileWord {
 		sep := regexp.QuoteMeta(string(os.PathSeparator))
 		wordRubout = fmt.Sprintf("%s[^%s]", sep, sep)
 		wordNext = fmt.Sprintf("[^%s]%s|(.$)", sep, sep)
 	}
-	spinner := []string{`⠋`, `⠙`, `⠹`, `⠸`, `⠼`, `⠴`, `⠦`, `⠧`, `⠇`, `⠏`}
-	if !opts.Unicode {
-		spinner = []string{`-`, `\`, `|`, `/`, `-`, `\`, `|`, `/`}
-	}
 	t := Terminal{
 		initDelay:   delay,
 		infoStyle:   opts.InfoStyle,
-		spinner:     spinner,
+		spinner:     makeSpinner(opts.Unicode),
 		queryLen:    [2]int{0, 0},
 		layout:      opts.Layout,
 		fullscreen:  fullscreen,
@@ -440,7 +488,8 @@ func NewTerminal(opts *Options, eventBox *util.EventBox) *Terminal {
 		selected:    make(map[int32]selectedItem),
 		reqBox:      util.NewEventBox(),
 		preview:     opts.Preview,
-		previewer:   previewer{"", 0, 0, previewBox != nil && !opts.Preview.hidden, false},
+		previewer:   previewer{0, []string{}, 0, previewBox != nil && !opts.Preview.hidden, false, true, ""},
+		previewed:   previewed{0, 0, 0, false},
 		previewBox:  previewBox,
 		eventBox:    eventBox,
 		mutex:       sync.Mutex{},
@@ -451,7 +500,7 @@ func NewTerminal(opts *Options, eventBox *util.EventBox) *Terminal {
 		killChan:    make(chan int),
 		tui:         renderer,
 		initFunc:    func() { renderer.Init() }}
-	t.prompt, t.promptLen = t.processTabs([]rune(opts.Prompt), 0)
+	t.prompt, t.promptLen = t.parsePrompt(opts.Prompt)
 	t.pointer, t.pointerLen = t.processTabs([]rune(opts.Pointer), 0)
 	t.marker, t.markerLen = t.processTabs([]rune(opts.Marker), 0)
 	// Pre-calculated empty pointer and marker signs
@@ -459,6 +508,19 @@ func NewTerminal(opts *Options, eventBox *util.EventBox) *Terminal {
 	t.markerEmpty = strings.Repeat(" ", t.markerLen)
 
 	return &t
+}
+
+func (t *Terminal) parsePrompt(prompt string) (func(), int) {
+	var state *ansiState
+	trimmed, colors, _ := extractColor(prompt, state, nil)
+	item := &Item{text: util.ToChars([]byte(trimmed)), colors: colors}
+	output := func() {
+		t.printHighlighted(
+			Result{item: item}, t.strong, tui.ColPrompt, tui.ColPrompt, false, false)
+	}
+	_, promptLen := t.processTabs([]rune(trimmed), 0)
+
+	return output, promptLen
 }
 
 func (t *Terminal) noInfoLine() bool {
@@ -574,12 +636,12 @@ const (
 	maxDisplayWidthCalc = 1024
 )
 
-func calculateSize(base int, size sizeSpec, margin int, minSize int) int {
-	max := base - margin
+func calculateSize(base int, size sizeSpec, occupied int, minSize int, pad int) int {
+	max := base - occupied
 	if size.percent {
 		return util.Constrain(int(float64(base)*0.01*size.size), minSize, max)
 	}
-	return util.Constrain(int(size.size), minSize, max)
+	return util.Constrain(int(size.size)+pad, minSize, max)
 }
 
 func (t *Terminal) resizeWindows() {
@@ -638,8 +700,12 @@ func (t *Terminal) resizeWindows() {
 	}
 	if t.pborder != nil {
 		t.pborder.Close()
+	}
+	if t.pwindow != nil {
 		t.pwindow.Close()
 	}
+	// Reset preview version so that full redraw occurs
+	t.previewed.version = 0
 
 	width := screenWidth - marginInt[1] - marginInt[3]
 	height := screenHeight - marginInt[0] - marginInt[2]
@@ -662,38 +728,45 @@ func (t *Terminal) resizeWindows() {
 	noBorder := tui.MakeBorderStyle(tui.BorderNone, t.unicode)
 	if previewVisible {
 		createPreviewWindow := func(y int, x int, w int, h int) {
-			previewBorder := tui.MakeBorderStyle(tui.BorderRounded, t.unicode)
-			if !t.preview.border {
-				previewBorder = tui.MakeTransparentBorder()
+			pwidth := w
+			pheight := h
+			if t.preview.border != tui.BorderNone {
+				previewBorder := tui.MakeBorderStyle(t.preview.border, t.unicode)
+				t.pborder = t.tui.NewWindow(y, x, w, h, true, previewBorder)
+				pwidth -= 4
+				pheight -= 2
+				x += 2
+				y += 1
+			} else {
+				previewBorder := tui.MakeTransparentBorder()
+				t.pborder = t.tui.NewWindow(y, x, w, h, true, previewBorder)
+				pwidth -= 4
+				x += 2
 			}
-			t.pborder = t.tui.NewWindow(y, x, w, h, true, previewBorder)
-			pwidth := w - 4
-			// ncurses auto-wraps the line when the cursor reaches the right-end of
-			// the window. To prevent unintended line-wraps, we use the width one
-			// column larger than the desired value.
-			if !t.preview.wrap && t.tui.DoesAutoWrap() {
-				pwidth += 1
-			}
-			t.pwindow = t.tui.NewWindow(y+1, x+2, pwidth, h-2, true, noBorder)
+			t.pwindow = t.tui.NewWindow(y, x, pwidth, pheight, true, noBorder)
+		}
+		verticalPad := 2
+		if t.preview.border == tui.BorderNone {
+			verticalPad = 0
 		}
 		switch t.preview.position {
 		case posUp:
-			pheight := calculateSize(height, t.preview.size, minHeight, 3)
+			pheight := calculateSize(height, t.preview.size, minHeight, 3, verticalPad)
 			t.window = t.tui.NewWindow(
 				marginInt[0]+pheight, marginInt[3], width, height-pheight, false, noBorder)
 			createPreviewWindow(marginInt[0], marginInt[3], width, pheight)
 		case posDown:
-			pheight := calculateSize(height, t.preview.size, minHeight, 3)
+			pheight := calculateSize(height, t.preview.size, minHeight, 3, verticalPad)
 			t.window = t.tui.NewWindow(
 				marginInt[0], marginInt[3], width, height-pheight, false, noBorder)
 			createPreviewWindow(marginInt[0]+height-pheight, marginInt[3], width, pheight)
 		case posLeft:
-			pwidth := calculateSize(width, t.preview.size, minWidth, 5)
+			pwidth := calculateSize(width, t.preview.size, minWidth, 5, 4)
 			t.window = t.tui.NewWindow(
 				marginInt[0], marginInt[3]+pwidth, width-pwidth, height, false, noBorder)
 			createPreviewWindow(marginInt[0], marginInt[3], pwidth, height)
 		case posRight:
-			pwidth := calculateSize(width, t.preview.size, minWidth, 5)
+			pwidth := calculateSize(width, t.preview.size, minWidth, 5, 4)
 			t.window = t.tui.NewWindow(
 				marginInt[0], marginInt[3], width-pwidth, height, false, noBorder)
 			createPreviewWindow(marginInt[0], marginInt[3]+width-pwidth, pwidth, height)
@@ -762,11 +835,19 @@ func (t *Terminal) placeCursor() {
 
 func (t *Terminal) printPrompt() {
 	t.move(0, 0, true)
-	t.window.CPrint(tui.ColPrompt, t.strong, t.prompt)
+	t.prompt()
 
 	before, after := t.updatePromptOffset()
 	t.window.CPrint(tui.ColNormal, t.strong, string(before))
 	t.window.CPrint(tui.ColNormal, t.strong, string(after))
+}
+
+func (t *Terminal) trimMessage(message string, maxWidth int) string {
+	if len(message) <= maxWidth {
+		return message
+	}
+	runes, _ := t.trimRight([]rune(message), maxWidth-2)
+	return string(runes) + strings.Repeat(".", util.Constrain(maxWidth, 0, 2))
 }
 
 func (t *Terminal) printInfo() {
@@ -820,11 +901,7 @@ func (t *Terminal) printInfo() {
 	if t.failed != nil && t.count == 0 {
 		output = fmt.Sprintf("[Command failed: %s]", *t.failed)
 	}
-	maxWidth := t.window.Width() - pos
-	if len(output) > maxWidth {
-		outputRunes, _ := t.trimRight([]rune(output), maxWidth-2)
-		output = string(outputRunes) + strings.Repeat(".", util.Constrain(maxWidth, 0, 2))
-	}
+	output = t.trimMessage(output, t.window.Width()-pos)
 	t.window.CPrint(tui.ColInfo, 0, output)
 }
 
@@ -999,7 +1076,7 @@ func (t *Terminal) printHighlighted(result Result, attr tui.Attr, col1 tui.Color
 	}
 
 	offsets := result.colorOffsets(charOffsets, t.theme, col2, attr, current)
-	maxWidth := t.window.Width() - 3
+	maxWidth := t.window.Width() - (t.pointerLen + t.markerLen + 1)
 	maxe = util.Constrain(maxe+util.Min(maxWidth/2-2, t.hscrollOff), 0, len(text))
 	displayWidth := t.displayWidthWithLimit(text, 0, maxWidth)
 	if displayWidth > maxWidth {
@@ -1075,28 +1152,47 @@ func (t *Terminal) printHighlighted(result Result, attr tui.Attr, col1 tui.Color
 	return displayWidth
 }
 
-func (t *Terminal) printPreview() {
-	if !t.hasPreviewWindow() {
-		return
+func (t *Terminal) renderPreviewSpinner() {
+	numLines := len(t.previewer.lines)
+	spin := t.previewer.spinner
+	if len(spin) > 0 || t.previewer.scrollable {
+		maxWidth := t.pwindow.Width()
+		if !t.previewer.scrollable {
+			if maxWidth > 0 {
+				t.pwindow.Move(0, maxWidth-1)
+				t.pwindow.CPrint(tui.ColSpinner, t.strong, spin)
+			}
+		} else {
+			offsetString := fmt.Sprintf("%d/%d", t.previewer.offset+1, numLines)
+			if len(spin) > 0 {
+				spin += " "
+				maxWidth -= 2
+			}
+			offsetRunes, _ := t.trimRight([]rune(offsetString), maxWidth)
+			pos := maxWidth - t.displayWidth(offsetRunes)
+			t.pwindow.Move(0, pos)
+			if maxWidth > 0 {
+				t.pwindow.CPrint(tui.ColSpinner, t.strong, spin)
+				t.pwindow.CPrint(tui.ColInfo, tui.Reverse, string(offsetRunes))
+			}
+		}
 	}
-	t.pwindow.Erase()
+}
 
+func (t *Terminal) renderPreviewText(unchanged bool) {
 	maxWidth := t.pwindow.Width()
-	if t.tui.DoesAutoWrap() {
-		maxWidth -= 1
-	}
-	reader := bufio.NewReader(strings.NewReader(t.previewer.text))
 	lineNo := -t.previewer.offset
 	height := t.pwindow.Height()
-	t.previewer.more = t.previewer.offset > 0
+	if unchanged {
+		t.pwindow.Move(0, 0)
+	} else {
+		t.previewed.filled = false
+		t.pwindow.Erase()
+	}
 	var ansi *ansiState
-	for ; ; lineNo++ {
-		line, err := reader.ReadString('\n')
-		eof := err == io.EOF
-		if !eof {
-			line = line[:len(line)-1]
-		}
+	for _, line := range t.previewer.lines {
 		if lineNo >= height || t.pwindow.Y() == height-1 && t.pwindow.X() > 0 {
+			t.previewed.filled = true
 			break
 		} else if lineNo >= 0 {
 			var fillRet tui.FillReturn
@@ -1115,29 +1211,53 @@ func (t *Terminal) printPreview() {
 				}
 				return fillRet == tui.FillContinue
 			})
-			t.previewer.more = t.previewer.more || t.pwindow.Y() == height-1 && t.pwindow.X() == t.pwindow.Width()
+			t.previewer.scrollable = t.previewer.scrollable || t.pwindow.Y() == height-1 && t.pwindow.X() == t.pwindow.Width()
 			if fillRet == tui.FillNextLine {
 				continue
 			} else if fillRet == tui.FillSuspend {
+				t.previewed.filled = true
 				break
 			}
-			t.pwindow.Fill("\n")
+			if unchanged && lineNo == 0 {
+				break
+			}
 		}
-		if eof {
-			break
-		}
+		lineNo++
 	}
-	t.pwindow.FinishFill()
-	if t.previewer.lines > height {
-		t.previewer.more = true
-		offset := fmt.Sprintf("%d/%d", t.previewer.offset+1, t.previewer.lines)
-		pos := t.pwindow.Width() - len(offset)
-		if t.tui.DoesAutoWrap() {
-			pos -= 1
-		}
-		t.pwindow.Move(0, pos)
-		t.pwindow.CPrint(tui.ColInfo, tui.Reverse, offset)
+	if !unchanged {
+		t.pwindow.FinishFill()
 	}
+}
+
+func (t *Terminal) printPreview() {
+	if !t.hasPreviewWindow() {
+		return
+	}
+	numLines := len(t.previewer.lines)
+	height := t.pwindow.Height()
+	unchanged := (t.previewed.filled || numLines == t.previewed.numLines) &&
+		t.previewer.version == t.previewed.version &&
+		t.previewer.offset == t.previewed.offset
+	t.previewer.scrollable = t.previewer.offset > 0 || numLines > height
+	t.renderPreviewText(unchanged)
+	t.renderPreviewSpinner()
+	t.previewed.numLines = numLines
+	t.previewed.version = t.previewer.version
+	t.previewed.offset = t.previewer.offset
+}
+
+func (t *Terminal) printPreviewDelayed() {
+	if !t.hasPreviewWindow() || len(t.previewer.lines) > 0 && t.previewed.version == t.previewer.version {
+		return
+	}
+
+	t.previewer.scrollable = false
+	t.renderPreviewText(true)
+
+	message := t.trimMessage("Loading ..", t.pwindow.Width())
+	pos := t.pwindow.Width() - len(message)
+	t.pwindow.Move(0, pos)
+	t.pwindow.CPrint(tui.ColInfo, tui.Reverse, message)
 }
 
 func (t *Terminal) processTabs(runes []rune, prefixWidth int) (string, int) {
@@ -1172,7 +1292,10 @@ func (t *Terminal) refresh() {
 			windows = append(windows, t.border)
 		}
 		if t.hasPreviewWindow() {
-			windows = append(windows, t.pborder, t.pwindow)
+			if t.pborder != nil {
+				windows = append(windows, t.pborder)
+			}
+			windows = append(windows, t.pwindow)
 		}
 		windows = append(windows, t.window)
 		t.tui.RefreshWindows(windows)
@@ -1196,7 +1319,8 @@ func findLastMatch(pattern string, str string) int {
 	if locs == nil {
 		return -1
 	}
-	return locs[len(locs)-1][0]
+	prefix := []rune(str[:locs[len(locs)-1][0]])
+	return len(prefix)
 }
 
 func findFirstMatch(pattern string, str string) int {
@@ -1208,7 +1332,8 @@ func findFirstMatch(pattern string, str string) int {
 	if loc == nil {
 		return -1
 	}
-	return loc[0]
+	prefix := []rune(str[:loc[0]])
+	return len(prefix)
 }
 
 func copySlice(slice []rune) []rune {
@@ -1316,6 +1441,50 @@ func cleanTemporaryFiles() {
 	activeTempFiles = []string{}
 }
 
+func (t *Terminal) replacePlaceholder(template string, forcePlus bool, input string, list []*Item) string {
+	return replacePlaceholder(
+		template, t.ansi, t.delimiter, t.printsep, forcePlus, input, list)
+}
+
+// Ascii to positive integer
+func atopi(s string) int {
+	matches := numericPrefix.FindStringSubmatch(s)
+	if len(matches) < 2 {
+		return 0
+	}
+	n, e := strconv.Atoi(matches[1])
+	if e != nil || n < 1 {
+		return 0
+	}
+	return n
+}
+
+func (t *Terminal) evaluateScrollOffset(list []*Item, height int) int {
+	offsetExpr := t.replacePlaceholder(t.preview.scroll, false, "", list)
+	nums := strings.Split(offsetExpr, "-")
+	switch len(nums) {
+	case 0:
+		return 0
+	case 1, 2:
+		base := atopi(nums[0])
+		if base == 0 {
+			return 0
+		} else if len(nums) == 1 {
+			return base - 1
+		}
+		if nums[1][0] == '/' {
+			denom := atopi(nums[1][1:])
+			if denom == 0 {
+				return base
+			}
+			return base - height/denom
+		}
+		return base - atopi(nums[1]) - 1
+	default:
+		return 0
+	}
+}
+
 func replacePlaceholder(template string, stripAnsi bool, delimiter Delimiter, printsep string, forcePlus bool, query string, allItems []*Item) string {
 	current := allItems[:1]
 	selected := allItems[1:]
@@ -1414,7 +1583,7 @@ func (t *Terminal) executeCommand(template string, forcePlus bool, background bo
 	if !valid {
 		return
 	}
-	command := replacePlaceholder(template, t.ansi, t.delimiter, t.printsep, forcePlus, string(t.input), list)
+	command := t.replacePlaceholder(template, forcePlus, string(t.input), list)
 	cmd := util.ExecCommand(command, false)
 	if !background {
 		cmd.Stdin = os.Stdin
@@ -1422,13 +1591,13 @@ func (t *Terminal) executeCommand(template string, forcePlus bool, background bo
 		cmd.Stderr = os.Stderr
 		t.tui.Pause(true)
 		cmd.Run()
-		t.tui.Resume(true)
+		t.tui.Resume(true, false)
 		t.redraw()
 		t.refresh()
 	} else {
 		t.tui.Pause(false)
 		cmd.Run()
-		t.tui.Resume(false)
+		t.tui.Resume(false, false)
 	}
 	cleanTemporaryFiles()
 }
@@ -1455,8 +1624,8 @@ func (t *Terminal) currentItem() *Item {
 
 func (t *Terminal) buildPlusList(template string, forcePlus bool) (bool, []*Item) {
 	current := t.currentItem()
-	_, plus, query := hasPreviewFlags(template)
-	if !(query && len(t.input) > 0 || (forcePlus || plus) && len(t.selected) > 0) {
+	slot, plus, query := hasPreviewFlags(template)
+	if !(!slot || query && len(t.input) > 0 || (forcePlus || plus) && len(t.selected) > 0) {
 		return current != nil, []*Item{current, current}
 	}
 
@@ -1582,68 +1751,141 @@ func (t *Terminal) Loop() {
 
 	if t.hasPreviewer() {
 		go func() {
+			version := 0
 			for {
-				var request []*Item
+				var items []*Item
+				var commandTemplate string
+				var pwindow tui.Window
 				t.previewBox.Wait(func(events *util.Events) {
 					for req, value := range *events {
 						switch req {
 						case reqPreviewEnqueue:
-							request = value.([]*Item)
+							request := value.(previewRequest)
+							commandTemplate = request.template
+							items = request.list
+							pwindow = request.pwindow
 						}
 					}
 					events.Clear()
 				})
+				version++
 				// We don't display preview window if no match
-				if request[0] != nil {
-					command := replacePlaceholder(t.preview.command,
-						t.ansi, t.delimiter, t.printsep, false, string(t.Input()), request)
+				if items[0] != nil {
+					command := t.replacePlaceholder(commandTemplate, false, string(t.Input()), items)
+					initialOffset := 0
 					cmd := util.ExecCommand(command, true)
-					if t.pwindow != nil {
+					if pwindow != nil {
+						height := pwindow.Height()
+						initialOffset = util.Max(0, t.evaluateScrollOffset(items, height))
 						env := os.Environ()
-						lines := fmt.Sprintf("LINES=%d", t.pwindow.Height())
-						columns := fmt.Sprintf("COLUMNS=%d", t.pwindow.Width())
+						lines := fmt.Sprintf("LINES=%d", height)
+						columns := fmt.Sprintf("COLUMNS=%d", pwindow.Width())
 						env = append(env, lines)
 						env = append(env, "FZF_PREVIEW_"+lines)
 						env = append(env, columns)
 						env = append(env, "FZF_PREVIEW_"+columns)
 						cmd.Env = env
 					}
-					var out bytes.Buffer
-					cmd.Stdout = &out
-					cmd.Stderr = &out
-					err := cmd.Start()
-					if err != nil {
-						out.Write([]byte(err.Error()))
-					}
+
+					out, _ := cmd.StdoutPipe()
+					cmd.Stderr = cmd.Stdout
+					reader := bufio.NewReader(out)
+					eofChan := make(chan bool)
 					finishChan := make(chan bool, 1)
-					updateChan := make(chan bool)
-					go func() {
-						select {
-						case code := <-t.killChan:
-							if code != exitCancel {
-								util.KillCommand(cmd)
-								os.Exit(code)
-							} else {
-								select {
-								case <-time.After(previewCancelWait):
-									util.KillCommand(cmd)
-									updateChan <- true
-								case <-finishChan:
-									updateChan <- false
+					reapChan := make(chan bool)
+					err := cmd.Start()
+					reaps := 0
+					if err != nil {
+						t.reqBox.Set(reqPreviewDisplay, previewResult{version, []string{err.Error()}, 0, ""})
+					} else {
+						reaps = 2
+						lineChan := make(chan eachLine)
+						// Goroutine 1 reads process output
+						go func() {
+							for {
+								line, err := reader.ReadString('\n')
+								lineChan <- eachLine{line, err}
+								if err != nil {
+									break
 								}
 							}
-						case <-finishChan:
-							updateChan <- false
+							eofChan <- true
+						}()
+						// Goroutine 2 periodically requests rendering
+						go func(version int) {
+							lines := []string{}
+							spinner := makeSpinner(t.unicode)
+							spinnerIndex := -1 // Delay initial rendering by an extra tick
+							ticker := time.NewTicker(previewChunkDelay)
+							offset := initialOffset
+						Loop:
+							for {
+								select {
+								case <-ticker.C:
+									if len(lines) > 0 && len(lines) >= initialOffset {
+										if spinnerIndex >= 0 {
+											spin := spinner[spinnerIndex%len(spinner)]
+											t.reqBox.Set(reqPreviewDisplay, previewResult{version, lines, offset, spin})
+											offset = -1
+										}
+										spinnerIndex++
+									}
+								case eachLine := <-lineChan:
+									line := eachLine.line
+									err := eachLine.err
+									if len(line) > 0 {
+										lines = append(lines, line)
+									}
+									if err != nil {
+										if len(lines) > 0 {
+											t.reqBox.Set(reqPreviewDisplay, previewResult{version, lines, offset, ""})
+										}
+										break Loop
+									}
+								}
+							}
+							ticker.Stop()
+							reapChan <- true
+						}(version)
+					}
+					// Goroutine 3 is responsible for cancelling running preview command
+					go func(version int) {
+						timer := time.NewTimer(previewDelayed)
+					Loop:
+						for {
+							select {
+							case <-timer.C:
+								t.reqBox.Set(reqPreviewDelayed, version)
+							case code := <-t.killChan:
+								if code != exitCancel {
+									util.KillCommand(cmd)
+									os.Exit(code)
+								} else {
+									timer := time.NewTimer(previewCancelWait)
+									select {
+									case <-timer.C:
+										util.KillCommand(cmd)
+									case <-finishChan:
+									}
+									timer.Stop()
+								}
+								break Loop
+							case <-finishChan:
+								break Loop
+							}
 						}
-					}()
-					cmd.Wait()
+						timer.Stop()
+						reapChan <- true
+					}(version)
+					<-eofChan
+					cmd.Wait() // NOTE: We should not call Wait before EOF
 					finishChan <- true
-					if out.Len() > 0 || !<-updateChan {
-						t.reqBox.Set(reqPreviewDisplay, out.String())
+					for i := 0; i < reaps; i++ {
+						<-reapChan
 					}
 					cleanTemporaryFiles()
 				} else {
-					t.reqBox.Set(reqPreviewDisplay, "")
+					t.reqBox.Set(reqPreviewDisplay, previewResult{version, nil, 0, ""})
 				}
 			}
 		}()
@@ -1657,6 +1899,14 @@ func (t *Terminal) Loop() {
 		}
 		// prof.Stop()
 		t.killPreview(code)
+	}
+
+	refreshPreview := func(command string) {
+		if len(command) > 0 && t.isPreviewEnabled() {
+			_, list := t.buildPlusList(command, false)
+			t.cancelPreview()
+			t.previewBox.Set(reqPreviewEnqueue, previewRequest{command, t.pwindow, list})
+		}
 	}
 
 	go func() {
@@ -1685,11 +1935,7 @@ func (t *Terminal) Loop() {
 						if focusedIndex != currentIndex || version != t.version {
 							version = t.version
 							focusedIndex = currentIndex
-							if t.isPreviewEnabled() {
-								_, list := t.buildPlusList(t.preview.command, false)
-								t.cancelPreview()
-								t.previewBox.Set(reqPreviewEnqueue, list)
-							}
+							refreshPreview(t.preview.command)
 						}
 					case reqJump:
 						if t.merger.Length() == 0 {
@@ -1701,7 +1947,7 @@ func (t *Terminal) Loop() {
 					case reqRefresh:
 						t.suppress = false
 					case reqReinit:
-						t.tui.Resume(t.fullscreen)
+						t.tui.Resume(t.fullscreen, true)
 						t.redraw()
 					case reqRedraw:
 						t.redraw()
@@ -1713,12 +1959,19 @@ func (t *Terminal) Loop() {
 							return exitNoMatch
 						})
 					case reqPreviewDisplay:
-						t.previewer.text = value.(string)
-						t.previewer.lines = strings.Count(t.previewer.text, "\n")
-						t.previewer.offset = 0
+						result := value.(previewResult)
+						t.previewer.version = result.version
+						t.previewer.lines = result.lines
+						t.previewer.spinner = result.spinner
+						if result.offset >= 0 {
+							t.previewer.offset = util.Constrain(result.offset, 0, len(t.previewer.lines)-1)
+						}
 						t.printPreview()
 					case reqPreviewRefresh:
 						t.printPreview()
+					case reqPreviewDelayed:
+						t.previewer.version = value.(int)
+						t.printPreviewDelayed()
 					case reqPrintQuery:
 						exit(func() int {
 							t.printer(string(t.input))
@@ -1738,6 +1991,7 @@ func (t *Terminal) Loop() {
 	for looping {
 		var newCommand *string
 		changed := false
+		beof := false
 		queryChanged := false
 
 		event := t.tui.GetChar()
@@ -1754,6 +2008,14 @@ func (t *Terminal) Loop() {
 				}
 			}
 		}
+		togglePreview := func(enabled bool) {
+			if t.previewer.enabled != enabled {
+				t.previewer.enabled = enabled
+				t.tui.Clear()
+				t.resizeWindows()
+				req(reqPrompt, reqList, reqInfo, reqHeader)
+			}
+		}
 		toggle := func() bool {
 			if t.cy < t.merger.Length() && t.toggleItem(t.merger.Get(t.cy).item) {
 				req(reqInfo)
@@ -1762,11 +2024,15 @@ func (t *Terminal) Loop() {
 			return false
 		}
 		scrollPreview := func(amount int) {
-			if !t.previewer.more {
+			if !t.previewer.scrollable {
 				return
 			}
-			newOffset := util.Constrain(
-				t.previewer.offset+amount, 0, t.previewer.lines-1)
+			newOffset := t.previewer.offset + amount
+			numLines := len(t.previewer.lines)
+			if t.preview.cycle {
+				newOffset = (newOffset + numLines) % numLines
+			}
+			newOffset = util.Constrain(newOffset, 0, numLines-1)
 			if t.previewer.offset != newOffset {
 				t.previewer.offset = newOffset
 				req(reqPreviewRefresh)
@@ -1802,17 +2068,15 @@ func (t *Terminal) Loop() {
 				return false
 			case actTogglePreview:
 				if t.hasPreviewer() {
-					t.previewer.enabled = !t.previewer.enabled
-					t.tui.Clear()
-					t.resizeWindows()
+					togglePreview(!t.previewer.enabled)
 					if t.previewer.enabled {
 						valid, list := t.buildPlusList(t.preview.command, false)
 						if valid {
 							t.cancelPreview()
-							t.previewBox.Set(reqPreviewEnqueue, list)
+							t.previewBox.Set(reqPreviewEnqueue,
+								previewRequest{t.preview.command, t.pwindow, list})
 						}
 					}
-					req(reqPrompt, reqList, reqInfo, reqHeader)
 				}
 			case actTogglePreviewWrap:
 				if t.hasPreviewWindow() {
@@ -1838,6 +2102,14 @@ func (t *Terminal) Loop() {
 				if t.hasPreviewWindow() {
 					scrollPreview(t.pwindow.Height())
 				}
+			case actPreviewHalfPageUp:
+				if t.hasPreviewWindow() {
+					scrollPreview(-t.pwindow.Height() / 2)
+				}
+			case actPreviewHalfPageDown:
+				if t.hasPreviewWindow() {
+					scrollPreview(t.pwindow.Height() / 2)
+				}
 			case actBeginningOfLine:
 				t.cx = 0
 			case actBackwardChar:
@@ -1846,6 +2118,11 @@ func (t *Terminal) Loop() {
 				}
 			case actPrintQuery:
 				req(reqPrintQuery)
+			case actPreview:
+				togglePreview(true)
+				refreshPreview(a.a)
+			case actRefreshPreview:
+				refreshPreview(t.preview.command)
 			case actReplaceQuery:
 				if t.cy >= 0 && t.cy < t.merger.Length() {
 					t.input = t.merger.Get(t.cy).item.text.ToRunes()
@@ -1881,6 +2158,7 @@ func (t *Terminal) Loop() {
 					t.cx++
 				}
 			case actBackwardDeleteChar:
+				beof = len(t.input) == 0
 				if t.cx > 0 {
 					t.input = append(t.input[:t.cx-1], t.input[t.cx:]...)
 					t.cx--
@@ -1973,16 +2251,19 @@ func (t *Terminal) Loop() {
 				t.vset(0)
 				req(reqList)
 			case actUnixLineDiscard:
+				beof = len(t.input) == 0
 				if t.cx > 0 {
 					t.yanked = copySlice(t.input[:t.cx])
 					t.input = t.input[t.cx:]
 					t.cx = 0
 				}
 			case actUnixWordRubout:
+				beof = len(t.input) == 0
 				if t.cx > 0 {
 					t.rubout("\\s\\S")
 				}
 			case actBackwardKillWord:
+				beof = len(t.input) == 0
 				if t.cx > 0 {
 					t.rubout(t.wordRubout)
 				}
@@ -2118,8 +2399,7 @@ func (t *Terminal) Loop() {
 					valid = !slot || query
 				}
 				if valid {
-					command := replacePlaceholder(a.a,
-						t.ansi, t.delimiter, t.printsep, false, string(t.input), list)
+					command := t.replacePlaceholder(a.a, false, string(t.input), list)
 					newCommand = &command
 				}
 			}
@@ -2142,6 +2422,11 @@ func (t *Terminal) Loop() {
 			changed = changed || queryChanged
 			if onChanges, prs := t.keymap[tui.Change]; queryChanged && prs {
 				if !doActions(onChanges, tui.Change) {
+					continue
+				}
+			}
+			if onEOFs, prs := t.keymap[tui.BackwardEOF]; beof && prs {
+				if !doActions(onEOFs, tui.BackwardEOF) {
 					continue
 				}
 			}