+++ .vim_runtime/my_plugins/fzf/src/tui/dummy.go
@@ -25,17 +25,16 @@ const (
 	Reverse          = Attr(1 << 6)
 )
 
-func (r *FullscreenRenderer) Init()       {}
-func (r *FullscreenRenderer) Pause(bool)  {}
-func (r *FullscreenRenderer) Resume(bool) {}
-func (r *FullscreenRenderer) Clear()      {}
-func (r *FullscreenRenderer) Refresh()    {}
-func (r *FullscreenRenderer) Close()      {}
+func (r *FullscreenRenderer) Init()             {}
+func (r *FullscreenRenderer) Pause(bool)        {}
+func (r *FullscreenRenderer) Resume(bool, bool) {}
+func (r *FullscreenRenderer) Clear()            {}
+func (r *FullscreenRenderer) Refresh()          {}
+func (r *FullscreenRenderer) Close()            {}
 
-func (r *FullscreenRenderer) DoesAutoWrap() bool { return false }
-func (r *FullscreenRenderer) GetChar() Event     { return Event{} }
-func (r *FullscreenRenderer) MaxX() int          { return 0 }
-func (r *FullscreenRenderer) MaxY() int          { return 0 }
+func (r *FullscreenRenderer) GetChar() Event { return Event{} }
+func (r *FullscreenRenderer) MaxX() int      { return 0 }
+func (r *FullscreenRenderer) MaxY() int      { return 0 }
 
 func (r *FullscreenRenderer) RefreshWindows(windows []Window) {}